Changed int to float

+++ quiz/index.php
@@ -18,14 +18,14 @@ switch($action)
         create_survey(filter_input(INPUT_POST, 'userID'));
         $survey_id = get_survey(filter_input(INPUT_POST, 'userID'));
         $question = get_question(filter_input(INPUT_POST, 'questionID'));
-        $qID = (int)filter_input(INPUT_POST, 'questionID');
+        $qID = (float)filter_input(INPUT_POST, 'questionID');
         $progress = ($qID/24)*100;
         include('quiz.php');
         break;
     case 'back':
         $survey_id = get_survey_by_id(filter_input(INPUT_POST, 'survey_id'));
         $question = get_question(get_last_question(filter_input(INPUT_POST, 'survey_id'))['question_id']);
-        $qID = (int)filter_input(INPUT_POST, 'questionID');
+        $qID = (float)filter_input(INPUT_POST, 'questionID');
         $progress = ($qID/24)*100;
         include('quiz.php');
         break;
@@ -35,7 +35,7 @@ switch($action)
             $survey_id = filter_input(INPUT_POST, 'survey_id');
             $question_id = filter_input(INPUT_POST, 'question_id');
             $answer = filter_input(INPUT_POST, 'answer');
-            $qID = (int)filter_input(INPUT_POST, 'questionID');
+            $qID = (float)filter_input(INPUT_POST, 'questionID');
             $progress = ($qID/24)*100;
             if(empty(get_answer($survey_id, $question_id))) //Check if question has been answered yet
                 create_answer($survey_id, $question_id, $answer); //Post answer to DB
